Remove YogaContextProvider
