+++ src/components/page/Page.tsx
@@ -1,7 +1,5 @@
 import * as React from 'react';
 import { BaseNodeProps, ChildrenProps, LayoutProps, ExportProps } from '../../types';
-import { yogaHandler } from '../../yogaHandler';
-import { YogaContextProvider } from '../../hooks/useYogaLayout';
 
 export interface PageProps extends BaseNodeProps, ChildrenProps, ExportProps {
     yogaRef: any;
