+++ src/renderer.tsx
@@ -5,6 +5,7 @@ import * as createReconciler from './realm-adopted/react-reconciler';
 
 import { GroupsProcessor } from './renderers/group/groupsProcessor';
 import { PREGROUP_NODE_TYPE } from './renderers/group/pregroupNode';
+import { addRoot } from './yoga';
 
 const isReactFigmaNode = child => child.getPluginData && child.getPluginData('isReactFigmaNode');
 
@@ -101,6 +102,7 @@ export const render = async (jsx: any, rootNode) => {
         supportsHydration: true,
         appendChildToContainer: (parentNode, childNode) => {
             appendToContainer(parentNode, childNode);
+            addRoot(childNode);
         },
         insertInContainerBefore: () => {},
         removeChildFromContainer: () => {},
